Fixed issue that caused ServerConnection threads to abort unexpectedly
The reason behind this was because multiple threads were trying to access the same string of bytes recieved from the client. However one got there slightly before the other thread and as a result corrupted the string resulting in the thread aborting because it couldn't parse the corrupted string correctly.

It now by adding server components as simply message recievers, instead of message processors that exist on a thread.